win-capture: Always refetch game capture addresses
This is a fast/immediate solution to a possible bug with caching the DLL
versions for game capture hook addresses - may as well just reload game
capture hook addresses each time the program is run for the time being
just to be safe.  Load time will increase a little for the time being
but it's worth it to prevent any issues with game capture.